# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于Go和Gin的博客系统API文档",
        "title": "Bmtdblog API",
        "termsOfService": "http://localhost:8090",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8090",
            "email": "3138910969@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
...
        "/signin": {
            "get": {
                "description": "显示用户登录表单页面",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户登录页面",
                "responses": {
                    "200": {
                        "description": "登录页面",
                        "schema": {
                            "type": "html"
                        }
                    }
                }
            },
            "post": {
                "description": "使用用户名和密码登录系统",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录失败，返回错误信息",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "302": {
                        "description": "登录成功，重定向到相应页面",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/signup": {
            "get": {
                "description": "显示用户注册表单页面",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户注册页面",
                "responses": {
                    "200": {
                        "description": "注册页面",
                        "schema": {
                            "type": "html"
                        }
                    }
                }
            },
            "post": {
                "description": "处理用户注册表单提交，创建新用户",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户注册提交",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱地址",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "手机号码",
                        "name": "telephone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "verifyCode",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册失败，返回注册页面并显示错误信息",
                        "schema": {
                            "type": "html"
                        }
                    },
                    "301": {
                        "description": "注册成功，重定向到登录页",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}